feat: Python scripts
